requirements.txt
uvicorn~=0.18.3
FastAPI~=0.85.0
starlette~=0.20.4
python-dotenv
python-multipart
geopandas
psycopg2
pygeos
